Fix binary output of undefined tiles
When a tile is undefined in binary HIST history output,
the value written out is zero, and not the undefined
value as set in the lis.config file.  netCDF and GRIB
output formats did not have this issue.  Essentally,
the code to set the undefined output values was copied
from these other formats and applied to binary format.

Resolves: #55