Bug fix for LDT downscaling bug when using GDAS
This fix is for the bug that occurred in LDT when using the 'Metforce temporal downscaling'
option along with the GDAS forcing dataset. The GDAS reader was not properly reseting and ingesting
new GDAS data at the downscaled time interval. A secondary flag was added to the reader that corrects
this issue. This fix is similar to the fix in the MERRA2 and CHIRPS readers in LDT.