Enables parallel support in HyMAP2
This commit enables HyMAP2 to be run in parallel. No
additional changes in the lis.config file are required.
The commit also enables the HyMAP2 restart files to be in
NetCDF.
Note that the parallel support does not
extend to the previous version of the HyMAP router or the
NLDAS router.
The following features in HyMAP2 are
currently not supported
1. Support for running with ensembles (The land model
can still be run in an ensemble mode. The ensemble average
is then used to perform the routing
2. Support for running with existing land model outputs
(The routing model only runs in online mode).
3. Support for binary restart files (only NetCDF output is
supported).

Resolves #7