Fix uninitialized variables in Noah-3.9 LSM over glacier
Two variables (flx4 and tsoil) were not defined over glacier
points, but were sent as LIS output variables.  This bug fix
sends the tsoil variable from the glacier points correctly to
the LIS output.  flx4 is not used as part of the glacier
physics, and is set to zero over glacier points.

Resolves: #110