Update documentation in lvt/configs directory
This commit updates documentation for the time series
location and metrics files in the lvt/configs directory.
In particular, it updates and details requirements for
how to specify these files based on recent code changes
within LVT.  Specific changes to these files include:

1) The METRICS file now must include two values for
   the threshold - first a minimum threshold value,
   and then a maximum threshold value.  Previously,
   only one value (one column) was required for the
   threshold.  Note that the threshold values are
   only used for categorical metrics; for the other
   metrics, undefined values of -9999.0 can be used.

2) The TS_LOCATIONS file has changed from using a
   minimum number of observations to a minimum fraction
   of the domain before temporal calculations will occur.

The file lvt/core/LVT_TSMod.F90 was also updated,
but mainly to copy the updated documentation into
the FORTRAN comments to be consistent with the
lvt/configs updates.  The one exception to changes
to the lvt/core/LVT_TSMod.F90 file was to change
the hard-coded value of the minimum fraction of
the domain for only location style 4.  Previous,
this was hard-coded to at least one valid value
within the polygon.  Users should be careful when
using location style 4, as they cannot specify a
minimum fraction of the domain that must have
valid observations in the TS_LOCATIONS file.

Resolves: #133