Fix post-2019Jun12 GDAS reader bug in LDT
This fix will allow the reader in LDT to read in f00 GDAS
data files after the 2019 June 12 12Z upgrade by NCEP.
The reader expected precipitation fields and averaged
radiation fields in the f00 GDAS data files. After the
upgrade, these fields were removed or were changed to
report instantaneous values, respectively. This fix is
similar to the LIS fix in pull request #160.

This also fixes a bug where 'glbdata_i' was not properly populated
with LDT_rc%udef.

The test files and configs can be found in:
/discover/nobackup/dpsarmie/work/gitLIS/LDT_GDASUpdateBugFix/ldt

Resolves: #167