Add support to compile LISF with GDAL and libgeotiff on conrad
The linker on conrad could not find libtiff.a, libjbig.a, and liblmza.a
and setting the LD_LIBRRAY_PATH environment variable did not help.  So
I added support to the Config.pl scripts to process LDT_LIBTIFF,
LDT_LIBJBIG, and LDT_LIBLZMA environment variables.  These environment
variables must be set for Cray systems.  For systems using shared
libraries, these environment variables are not needed.

Resolves: #192