Fix SNODAS LVT log messages and big_endian data read
This pull request fixes two issues with log messages
in the SNODAS LVT reader.  The first is that the line
showing successful processing of the SWE dat file now
prints the correct file name.  The other is that LVT
will print a log message when it cannot find a SNODAS
file for which it is looking.

Also addressed in this pull request is to specify in
the open statement that the SNODAS dat binary files
are natively "big_endian".  Now LVT will produce the
same results when processing SNODAS when configured/
compiled with either big or little endian.

Resolves: #124