Adds the observation simulator runmode in LDT
A new runmode that enables the simulation of synthetic observations
is added to LDT. This feature is intented for facilitating
OSSE experiments. The code samples from a given Nature run and
samples the specified fields, applies temporal, spatial tranformations,
temporal masks, and applies errors. Finally the simulated
observations are written out to an external file

Resolves #220