Address Fortify Audit issue "APSC-DV-002590, V-70277"
From report:

Buffer Overflow: Format String

* Offender is sprint() function. The string being assigned to msg ends
up being larger than allowed for msg, it looks like. The length of msg,
len, is set to 30+strlen(caller)+1. The number of “guaranteed”
characters is 31: ERR: %s -- Cannot allocate memory. With the exception
of %s, this is 31 characters. Assuming the additional +1 is for some
null character, 30 should be changed to 31 (or larger, for safety).