Update LIS config documentation for MERRA2 reader
This commit adds documentation for an option to use 2m wind
fields when using the MERRA2 reader. The documentation was
not originally included when the commit was first made. This
option was meant to compliment a secondary reference
evapotranspiration calculation option that is available in
LVT.

There was no issue opened for this fix.