Address Fortify Audit issue "APSC-DV-002570, V-70273"
Issues "APSC-DV-002580, V-70275" and "APSC-DV-003235, V-70391" are the same.

From report:

Poor Error Handling: Empty Catch Block

Multiple uses of "pass" to ignore errors.  Put in some type of handling
or alerting when these happen. If they must be ignored, at least log
them so it is known that they are occurring.