Fix Metric Ouput Directory config check in LVT
The check to see if the 'Metrics output directory:' option was
successfully read in occurs after two other configuration options
are read in by the program, which results in a check on the wrong
option.

This fix just reorders the subroutine calls so that the check is
directly after the subroutine call that reads in the 'Metrics
output directory:' option.

Resolves: #374