Prevent LIS from writing ASCII statistics data when option is off
This commit will prevent LIS from writing ASCII statistics data
to an incorrect output file when this option is turned off in the
lis.config file.  Only when netCDF output is chosen, LIS will try
to write the statistics data into a dummy file in the LIS running
directory.  When the statistics file is turned off in the config
file, this bug fix now correctly keeps LIS from writing any of
the statistics data into this dummy file.

Resolves: #384