Added sanity checks for iceage vs icemask for times other than 12Z.
Previously iceage was only updated at 12Z.  However, if LDTSI was coldstarted
at a different time, it was possible for iceage to be set to missing while
icemask (and icecon) was set to valid values.  The reason is that LDTSI
will use SNODEP as the first guess when coldstarting, but mismatches between
the LDTSI and SNODEP land/sea masks can cause the ice fields to be initialized
as missing.

This problem originally fixed itself when LDTSI was cycled to 12Z, and
iceage was finally updated.  But for early analyses, this mismatch would
cause LVT to crash in LDTSI post mode.