557 post mode now uses Welford algorithm for standard deviation.
The previous, straight-forward calculation of standard deviation was prone
to run-off error, which could lead to negative variances and imaginary
standard deviation.  This new formulation uses the Welford algorithm which
is more resistent to loss of precision.  Note, however, that the algorithm
is only used when processing multiple ensemble members; if a single
deterministic run is processed, standard deviation is just set to missing.