Fix LIS soil moisture DA QC checks for Noah-MP-4.0.1 LSM
This commit fixes a bug in the soil moisture DA quality
control (QC) codes for the Noah-MP-4.0.1 LSM.  The code
was setting the value for all patches of variables SMCMAX
and SMCWLT to the same value.  This fix correctly sets
the values for all patches for these variables.

A review was done for Noah-MP-3.6 and Noah-3.X quality
control soil moisture DA codes, and it was determined
that no fixes are required to these other LSMs.

Resolves: #534