Remove SMAP(NASA) LIS DA QC check config entry
This commit removes the SMAP(NASA) LIS soil moisture DA QC
(quality control) check config entry.  After this commit,
the SMAP QC will always be on for this product.

Changes to the lis.config were also made, along with a small
tweak to lislog messages, and to the exact specification of
another SMAP(NASA) config entry.

Resolves: #551