Add new modulefile on discover to work around missing stack traces
Intel-MPI-based programs do not produce a stack trace.

Matt Thompson worked with Intel on this issue.  They determined that the cause
is presence of libmlx-fi.so in FI_PROVIDER_PATH.  Matt came up with a
work-around by creating a new directory containing a copy of the Intel MPI
library files except for libmlx-fi.so.

You may see a performance impact when running in this environment, so it
is recommended that you use this modulefile only when debugging.  Fortunately,
this modifies only your running environment, so you do not have to recompile
LISF when you load this modulefile.