Fix grid check in AWRAL reader
A grid check in the AWRAL metforcing reader incorrectly compared
ncol to gnr. The check would therefore evaluate to false even if
the forcing grid was the same as the model.

The logical operator ".neqv" is converted to ".ne" as the former
requires both operands to be of type LOGICAL while the operands
here are INTEGER(4). This change fixes a "Operands of logical
operator '.neqv.' at (1) are INTEGER(4)/INTEGER(4)" compiler
error.