Refine JULES-5.0 snow DA checks against attrib file maximums
This commit refines JULES-5.0 snow DA checks as recommended
by @yyoon, @yhkwon81, and @emkemp on the comments to this
pull request.

Instead of hard-coding a value above which snow DA is not
performed, the maximum values in the attribs file are used.

Also, consistency checks between the SWE and snow depth
after snow DA against the maximums are performed, using
the snow density.

The recommended JULES-5.0 SWE and snow depth maximums are
lowered to their previous values, and now are consistent
with the numbers used by Noah-3.9 and Noah-MP-4.0.1 LSMs.