Ensure that every item written to filelist_name ends with a newline
The last entry in filelist_name was missing a newline character.  It is
ok for all other uses of the create_filelist routine.  However, the SMOS
NRT NNL2 soil moisture reader calls this routine twice and concantenates
the results.  So the last line of file_01.tx and the first line of
file_02.txt would run together.