Commented out LVT_statsEntry fields in LVT_stats object.
The presence of these fields causes GFORTRAN to fail when compiling with
optimization, and to struggle when optimization is turned off.  It turns
out that these entries are not actually used by LVT -- the actual
LVT_statsEntry objects are created and organized as a linked list.  Commenting
the code out does not affect the output from LVT.