Replace and()/or() functions with .and./.or. logical operators
The and() and or() functions seem to be compiler-specific functions that perform either a bitwise comparison or a logical comparison of its two arguments.  However, they do not work consistently across compilers.

Use iand()/ior() to perform a bitwise comparison of two integer values.

Use .and./.or. to perform a logical comparison of two logical values.

Examples:

`ior(1,2)` yields 3.

`(.true. .or. .false.)` yields .true..

`((i==1) .or. (i==2))` yields .true., for i = 1.

Compiler error:

....
       elseif (or(hr.eq.CNRS_em_obs_struc(n)%mask_hr_d_lower,hr.eq.CNRS_em_obs_struc(n)%mask_hr_d_upper)) then
               ^
ftn-416 crayftn: ERROR CNRS_EM_OBS_SETUP, File = ../optUE/type/paramestim/obs/CNRS/CNRS_em_obsMod.F90, Line = 537, Column = 16
  The conditional expression must be type logical.
....

....
       if(and(nobs_sum.ge.15,platform_sum.ge.2)) then
          ^
ftn-416 crayftn: ERROR READ_CNRS_MPDI_EM_OBSDATA, File = ../optUE/type/paramestim/obs/CNRS/read_CNRSdata_MPDI.F90, Line = 210, Column = 11
  The conditional expression must be type logical.
....

The following output demonstrates the inconsistent behaviour of the generic and() and or() functions.  The GNU compiler behaves correctly.  The Intel compiler works, but it returns -1 for .true. and 0 for .false. for logical comparisons, instead of returning a logical value.  (The Intel compiler treats 0 as .false. and non-0 as .true..)  The Cray compiler, for logical comparisons, sometimes returns a logical value and sometimes returns an integer.

GNU Fortran compiler:

Setting a to .true. and b to .true.
a is  T ; b is  T
and(a,b) is  T (expect .true.)
or(a,b) is  T (expect .true.)
Setting a to .false. and b to .true.
a is  F ; b is  T
and(a,b) is  F (expect .false.)
or(a,b) is  T (expect .true.)
Setting i to 1 and j to 2
i is            1 ; j is            2
and(i,j) is            0 (expect 0)
or(i,j) is            3 (expect 3)
Setting i to 1 and j to 3
i is            1 ; j is            3
and(i,j) is            1 (expect 1)
or(i,j) is            3 (expect 3)

Intel Fortran compiler:

Setting a to .true. and b to .true.
a is  T ; b is  T
and(a,b) is           -1 (expect .true.)
or(a,b) is           -1 (expect .true.)
Setting a to .false. and b to .true.
a is  F ; b is  T
and(a,b) is            0 (expect .false.)
or(a,b) is           -1 (expect .true.)
Setting i to 1 and j to 2
i is            1 ; j is            2
and(i,j) is            0 (expect 0)
or(i,j) is            3 (expect 3)
Setting i to 1 and j to 3
i is            1 ; j is            3
and(i,j) is            1 (expect 1)
or(i,j) is            3 (expect 3)

Cray Fortran compiler:

Setting a to .true. and b to .true.
a is  T ; b is  T
and(a,b) is  T (expect .true.)
or(a,b) is  T (expect .true.)
Setting a to .false. and b to .true.
a is  F ; b is  T
and(a,b) is  0 (expect .false.)
or(a,b) is  1 (expect .true.)
Setting i to 1 and j to 2
i is  1 ; j is  2
and(i,j) is  0 (expect 0)
or(i,j) is  3 (expect 3)
Setting i to 1 and j to 3
i is  1 ; j is  3
and(i,j) is  1 (expect 1)
or(i,j) is  3 (expect 3)

Test code:

program andor
   implicit none

   logical :: a, b
   integer :: i, j

   print*,'Setting a to .true. and b to .true.'
   a = .true.; b = .true.
   print*,'a is ', a, '; b is ', b
   print*,'and(a,b) is ', and(a,b), '(expect .true.)'
   print*,'or(a,b) is ', or(a,b), '(expect .true.)'

   print*,'Setting a to .false. and b to .true.'
   a = .false.; b = .true.
   print*,'a is ', a, '; b is ', b
   print*,'and(a,b) is ', and(a,b), '(expect .false.)'
   print*,'or(a,b) is ', or(a,b), '(expect .true.)'

   print*,'Setting i to 1 and j to 2'
   i=1; j=2
   print*,'i is ', i, '; j is ', j
   print*,'and(i,j) is ', and(i,j), '(expect 0)'
   print*,'or(i,j) is ', or(i,j), '(expect 3)'

   print*,'Setting i to 1 and j to 3'
   i=1; j=3
   print*,'i is ', i, '; j is ', j
   print*,'and(i,j) is ', and(i,j), '(expect 1)'
   print*,'or(i,j) is ', or(i,j), '(expect 3)'

   ! Not allowed by Cray compiler
   ! Not allowed by GNU compiler
   ! Intel compiler produces: true, false, true.
   !do i = -1, 1
   !   print*,'i = ', i
   !   if ( i ) then
   !      print*,'true'
   !   else
   !      print*,'false'
   !   endif
   !enddo

end program andor