New files for running LIS to generate atmos forcing for S2S.
This is a slimmed down version of the LIS NRT AGRMET Ops configuration for
Noah 3.9.  No land DA is performed, and only a single ensemble member is used.
The scripts assume that a single-day run will be initialized at 00Z from a
restart, and will save GRIB1 output and 24-hr restart files for subsequent
use.