Tightened mean OMB threshholds for spd10m and t2m variables.
This reduces the scatter of OMB values in empirical semivariograms plotting
from 28 days of data, with semivariograms created from 29 Nov to 5 Dec 2021.
The drawback is the reduction in error correlation length scale, which is
a big severe.  This may be another argument for switching to inverse
exponential instead of Gaussian correlation functions.