Adds the support for the GDDP reader
Adds the support for the NEX GDDP climate data as forcing. The plugin
temporally disaggregates the daily GDDP data to subdaily resolution.

The code also updates includes updates to the parameter reading routines
so that in a multiprocessor mode, each process directly reads the subsetted
data slice. The older version of the code read the global data and subsetted it.

Resolves #989