Modify GEFS reader to support 0.25 deg data
This commit will add support for the 0.25 deg GEFS data. The reader modifications
will keep the 0.50 deg functinoality and add a config option to allow the user
to swtich between 0.50 deg and 0.25 deg GEFS data. There is a bug with the
reforecast data when reading in the GEFS wind fields and this will be fixed in a
subsequent fix.