Fix use of input LAI for Noah-MP-4.0.1 vegetation options
For Noah-MP-4.0.1 dynamic vegetation options 7, 8, or 9, the input
LAI was not being correctly used in the model physics.  This pull
request fixes this bug, and allows the use of input LAI or SAI in
the model physics.  The LIS Users' Guide was also updated to give
further details of the config entries that need to be provided for
all of the different options.

Options 1 to 6 for this physics option is not affected by this fix.

Resolves: #1011