Fix the wind field when using GEFS 0.25 Reforecast
The GEFS reforecsast data at 0.25 degrees introduced a notation
change for the wind fields. The bug has been fixed and the GEFS
reader will now properly read in the wind fields from the reforecast2
data at 0.25 degrees and 1.0 degrees. The operational mode of the reader
is not affected by this change and works properly.

This is the final fix for all currently known bugs in the GEFS 0.25 degree reader.