Begins updating Public Testcase from p7.3 to p7.4
Begins editing the Public Testcase Walkthough so that it directs
users to support/lisf-public-7.4 instead of support/lisf-public-7.3.
Gramatical errors and other minor referencing errors are also fixed.
To finalize this commit, changes need to be made to the public
testcase .html in the SETUP section to point users to 7.4 instead
of 7.3.