Add support to specify RPC library
The OS upgrade on HPC-11 changed the available low level RPC library.
The system no longer has a static version to compile into the HDF4
library.  You must now specify the RPC library in the LDFLAGS variable
(both when building the HDF4 library itself and when compiling LISF).

This commit adds LDT_RPC, LIS_RPC, and LVT_RPC environment variables.
If these environment variables are defined, then the build scripts add
"-ltirpc" to the LDFLAGS variable for HDF4.  If these environment
variables are not defined, then the build scripts function as before
(nothing new is added to the LDFLAGS variable for HDF4).