Address errors raised by GNU compilers
The GNU Fortran compiler raised errors when compiling source code that
included mpif.h (needed when compiling with gfortran while using the
Intel MPI library) or used the HDF-EOS2 library.

The workaround is to add `-fallow-argument-mismatch` to the gfortran
compiler flags.

Note: On some systems, this compiler option is already added by the mpif90
wrapper script.