Modified RAPID code to compile w/o PETSc.
Wrapper preprocessor statements have been added to activate dummy subroutines
and modules if PETSc is not selected but RAPID is to be compiled.  This
is consistent with how we handle code that expects HDF4, HDF5, and HDFEOS.

Compilation has been tested w/o PETSc with GNU and Intel compilers.
Compilation has also been tested w/ PETSc with the Intel compiler.

Runtime tests are pending.