Correct intent error
This patch fixes an intent(out) vs intent(inout) error that
resulted in glaciers melting.

In SFCDIF1_GLACIER, the variable FV was declared as intent(out).
Technically, this means that FV was uninitialized.  FV was initialized
for the first iteration (ITER=1), but it was not initialized for
subsequent calls (iterations 2 through 5).