New lis.config option - # of dims in latlon field
This is an addition to Sujay's PR 1179 that changes the LIS lat/lon field output from 2d to 1d.
The number of latlon dimensions is now an option in the lis.config file.  This was done because
this enables support for 1D and 2D output.  Additinally, this maintains 2D support for projections
that cannot be written with 1D.

The logic works as follows:

no entry in config, invalid entry in config, entry of 1D in config: 1D output requested
	check if projection supports 1D
		if yes: write with 1D
		if no: write with 2D

entry of 2D in config: 2D output requested
	write with 2D