Add support to specify RPC library
The OS upgrade on narwhal changed the available low level RPC library.
The system no longer has a static version to compile into the HDF4
library.  You must now specify the RPC library in the LDFLAGS variable
(both when building the HDF4 library itself and when compiling LISF).

This OS upgrade is also missing a symbolic link to libtirpc.so.
This commit tweaks the use of the LDT_RPC, LIS_RPC, and LVT_RPC
environment variables.

If these environment variables are defined, then they must be set to the
full file name of the tirpc library (e.g., /lib64/libtirpc.so.3). Then
the build scripts will add this library to the LDFLAGS variable for
HDF4.  If these environment variables are not defined, then the build
scripts function as before (nothing new is added to the LDFLAGS variable
for HDF4).

Also see commit 935d662588b56b5140165051d451518d0992f53a.