Updates to LVT AGRMET reader.
Minute is now added to the logic for creating a filename.  This is
necessary when LVT is comparing AGRMET to another dataset with time
interval of 30 minutes (e.g., IMERG).  Before, LVT would use '00' as the
minute and would erroneously read in AGRMET files valid at the top of
the hour at 0030, 0330, etc. Now, LVT will search for files at the bottom
of the hour, fail to find them, and gracefully mark the data values as
"missing" for that time.