Added new dupQC routine for future use.
NOTE: This new routine is not used (the original dupQC is), as we do
not have time to evaluate impact on results.  (Plus the WIGOS files are
still not quite ready for use.)  But by adding this new routine
(USAF_dupqc_new), it will be straightforward to test potential a
potential switch.