Revised alert filenames for USAFSI.
(1) Character lengths are increased to support whole subroutine names,
and to allow total file name to be up to 255 characters (avoid truncation).

(2) Added valid date/time (YYYYMMDDHH) of the USAFSI analysis that was
being produced when alert was issued.

Not on only does this aid assessment of what is causing a problem when
running USAFSI, but it also helps avoid the situation where alert
numbers become too big to represent in the filename.  That can trigger
an infinite loop (which still needs to be fixed).

This was manually applied from commit bcd3bd9e312b069399b594b28a0e4ba7712f6569