Add dynamic lapse rate usage and output to MERRA-2 metforcing
This commit adds the reading and application of pre-generated
hourly MERRA-2 lapse rate maps.  This optional config will use
these dynamic lapse rates in place of the static lapse rate
(of -6.5 K/km) from the "lapse-rate" topographic correction.

Sujay Kumar wrote the initial version of this code, and David
Mocko and Sujay worked together to fix some code issues, add
code fault tolerance, and add the ability to output dynamic
lapse rate on the output grid.

For now, this code should remain in the eis-freshwater2 branch,
and should not be merged into the main branch yet.