Changes to support WIGOS in autotuning.
Identical results are found for gage-nwp, rh2m, spd10m, and t2m, using
28 days of OBA files produced ending 1 Sep 2024.  (Pre-WIGOS OBA files
were generated during the LISF 7.6 FOC work.)

NOTE:  Differences are noted in the IMERG data counts in the WIGOS OBA
files.  I suspect this is due to changes in the IMERG file archives
between the FOC work (in Sep) and the WIGOS work (in Dec).  I want to
rerun the FOC OBA LIS version, but that must wait until after Christmas
Break.