Add MERRA-2 dynamic lapse rate double-sided cutoff
This commit adds the application of a double-sided cutoff to
pre-generated hourly MERRA-2 lapse rate maps. This optional
functionality will replace any lapse rate value that falls above
(below) the maximum (minimum) cutoff threshold with the cutoff
threshold value. The function is invoked by an optional config entry to
apply the double-sided cutoff. If this config entry is not present, the
cutoffs will not be applied. The commit also includes optional config
entries to specify the minimum and maximum cutoff thresholds in units
of K/m. If the config entry for the maximum (minimum) threshold is not
present, the default is set to +0.01 (-0.01) K/m.

Kristen Whitney and David Mocko worked together to modify the code with
this functionality. Kristen Whitney performed code testing.

For now, this code should remain in the eis-freshwater2 branch, and
should not be merged into the main branch yet.

Credit where credit is due:
@dmocko helped with the corresponding code modifications and
adjustments.

Resolves: #1653