Add CERES SWdown metforcing reader for NLDAS-3
This commit adds a metforcing reader for the CERES SWdown
product over the NLDAS-3 domain.  The product is binary
little_endian real*4 hourly at 4-km over the domain.

This metforcing reader will read in the product and enable
interpolation and slope-aspect corrections to the running
grid.  Primarily, it will be run on the 1-km NLDAS-3 grid.

Note that this metforcing reader should NOT be merged into
the main branch at this time.

Resolves: #1668