Fix to compile LIS with "WRF coupling: Off"
Note that both the GCE coupling runmode and the NUOPC coupling runmode
depend on code in the the WRF coupling runmode.  So to use either of those
runmodes, you must enable "WRF coupling", which is its default setting.

Note 1) that both the GCE coupling runmode and the NUOPC coupling runmode
are not enabled by default and 2) that both require special build commands.