Convert cropt%value to an integer
The Cray Fortran compiler raised this error:

    ftn-319 ftn: ERROR DEFINE_AC_COMPARTMENTS, File = ../params/AquaCrop/define_AC_compartments.F90, Line = 205, Column = 63
      A subscript must be a scalar integer expression.

cropt is type LDT_paramEntry.  The value element is a real-valued array.
cropt represents a crop type index.  I confirmed that this index is
an integer.  So I converted cropt%value to an integer to use as an index
into the csizes array.